requirements/base.txt
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.6
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h11==0.14.0
httptools==0.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
Pillow==10.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
rsa==4.9
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
zope.interface==6.0
